refactored support into components
