+++ src/app/(app)/(home)/contact/page.tsx
@@ -1,0 +1,5 @@
+function Page() {
+  return <div>Contact</div>;
+}
+
+export default Page;
